Move `logstream` fixture to the top-level conftest for all tests
